streamlit/requirements.txt
streamlit
requests
pandas
requests
fastapi
uvicorn
httpx
pydantic
beautifulsoup4
google.generativeai
load_dotenv
matplotlib
geopy